+++ index.html
@@ -49,6 +49,6 @@
 
       <div id="loaderRoot"></div>
     </div>
-    <script src="./scripts/renderer.js"></script>
+    <script src="./src/renderer.js"></script>
   </body>
 </html>
